// src/types/api/points/point-function.ts
export type ApiPointFunction =
  | 'parcel'
  | 'parcel_send'
  | 'parcel_collect'
  | 'parcel_reverse_return_send'
  | 'standard_letter_collect'
  | 'standard_letter_send'
  | 'allegro_parcel_collect'
  | 'allegro_parcel_send'
  | 'allegro_parcel_reverse_return_send'
  | 'allegro_letter_collect'
  | 'allegro_letter_send'
  | 'allegro_letter_reverse_return_send'
  | 'allegro_courier_collect'
  | 'allegro_courier_send'
  | 'allegro_courier_reverse_return_send'
  | 'standard_courier_collect'
  | 'standard_courier_send'
  | 'standard_courier_reverse_return_send'
  | 'air_on_airport'
  | 'air_outside_airport'
  | 'cool_parcel_collect'
  | 'laundry'
  | 'avizo';